cdk bootstrap options display
Changed the way the cdk bootstrap options where displayed as having --toolkit-bucket-name on the next line was creating customer confusion, i.e. customer thought that --bootstrap-bucket-name and --toolkit-bucket-name were 2 different bucket and since "[string]" was on the same line as --toolkit-bucket-name, they took this as the only explanation if what the parameter meant so they had no idea what this was for.

If this display method is not accepted, alternatively use tables to display these options

Further, customer needs docs to be more clearer on whether bootstrap creates the bucket.

Please include documentation on what security is added to the bucket when boostrap creates it as well as if any life cycle is applied.