Clarifies the Arrays.asList method
The list returned by Arrays.asList is Arrays$ArrayList, not ArrayList. Since one is unmodifiable and one is not, this can trip up people. It should also be noted that Java 8's docs says Arrays.asList returns List, not ArrayList[1]. So to make sure that the docs always smell of freshly baked quality, I propose this modest change.

[1] https://docs.oracle.com/javase/8/docs/api/java/util/Arrays.html