Remove preview note from blueprints documentation (#473)
## Description

**Documentation Change Type:**

- [ ] Typo/grammar fix
- [X] Clarification or minor improvement
- [ ] New example or code snippet
- [ ] New section or topic
- [ ] Major restructuring or enhancement

**Affected Documentation Page(s):**
- https://docs.aws.amazon.com/cdk/v2/guide/blueprints.html

## Description of Changes
The Blueprints aka Property Injection feature is live since more than 3
months now. We don't plan to introduce any major changes to this
feature. At this point, we feel confident that all CDK customers can
leverage this feature. This change is to remove the **"CDK Blueprints is
in preview release and is subject to change"** note banner from the
documentation page.

## Motivation and Context
The Blueprints aka Property Injection feature is live since more than 3
months now. We don't plan to introduce any major changes to this
feature. At this point, we feel confident that all CDK customers can
leverage this feature. This change is to remove the **"CDK Blueprints is
in preview release and is subject to change"** note banner from the
documentation page.

## How Has This Been Validated?
Change is small enough to be reviewed manually.

## Screenshots (if appropriate)
- https://docs.aws.amazon.com/cdk/v2/guide/blueprints.html

## Checklist

- [X] My changes follow the [AsciiDoc
syntax](./CONTRIBUTING.md#asciidoc-syntax-reference) guidelines
- [X] I have previewed my changes using GitHub's preview or a local
AsciiDoc renderer
- [X] My changes maintain consistent formatting with the rest of the
documentation
- [X] I have checked that all links work correctly

## Additional Information
<!-- Any additional information or context that might be helpful for
reviewers -->